Collect all conf helper methods in the application_helper
